+++ tpcctables.cc
@@ -1059,13 +1059,13 @@ void TPCCTables::HistoryToCSV(int64_t num_warehouses) {
     his_f.close();
 }
 
-int32_t TPCCTables::itemSize() {
+int64_t TPCCTables::itemSize() {
     int32_t ret = 0;
     for (Item &item: items_) ret += item.Size();
     return ret;
 }
 
-int32_t TPCCTables::warehouseSize(int64_t num_warehouses) {
+int64_t TPCCTables::warehouseSize(int64_t num_warehouses) {
     int32_t ret = 0;
     for (int32_t i = 1; i <= num_warehouses; ++i) {
         Warehouse *w = findWarehouse(i);
@@ -1075,8 +1075,8 @@ int32_t TPCCTables::warehouseSize(int64_t num_warehouses) {
     return ret;
 }
 
-int32_t TPCCTables::districtSize(int64_t num_warehouses) {
-    int32_t ret = 0;
+int64_t TPCCTables::districtSize(int64_t num_warehouses) {
+    int64_t ret = 0;
     for (int32_t i = 1; i <= num_warehouses; ++i) {
         for (int32_t j = 1; j <= District::NUM_PER_WAREHOUSE; ++j) {
             District *d = findDistrict(i, j);
@@ -1087,8 +1087,8 @@ int32_t TPCCTables::districtSize(int64_t num_warehouses) {
     return ret;
 }
 
-int32_t TPCCTables::stockSize(int64_t num_warehouses) {
-    int32_t ret = 0;
+int64_t TPCCTables::stockSize(int64_t num_warehouses) {
+    int64_t ret = 0;
     for (int32_t i = 1; i <= num_warehouses; ++i) {
         for (int32_t j = 1; j <= Stock::NUM_STOCK_PER_WAREHOUSE; ++j) {
             Stock *s = findStock(i, j);
@@ -1099,8 +1099,8 @@ int32_t TPCCTables::stockSize(int64_t num_warehouses) {
     return ret;
 }
 
-int32_t TPCCTables::customerSize(int64_t num_warehouses) {
-    int32_t ret = 0;
+int64_t TPCCTables::customerSize(int64_t num_warehouses) {
+    int64_t ret = 0;
     for (int32_t i = 1; i <= num_warehouses; ++i) {
         for (int32_t j = 1; j <= District::NUM_PER_WAREHOUSE; ++j) {
             for (int32_t k = 1; k <= Customer::NUM_PER_DISTRICT; ++k) {
@@ -1113,8 +1113,8 @@ int32_t TPCCTables::customerSize(int64_t num_warehouses) {
     return ret;
 }
 
-int32_t TPCCTables::orderSize(int64_t num_warehouses, int64_t num_transactions) {
-    int32_t ret = 0;
+int64_t TPCCTables::orderSize(int64_t num_warehouses, int64_t num_transactions) {
+    int64_t ret = 0;
     int32_t krange = Order::INITIAL_ORDERS_PER_DISTRICT + num_transactions;
     for (int32_t i = 1; i <= num_warehouses; ++i) {
         for (int32_t j = 1; j <= District::NUM_PER_WAREHOUSE; ++j) {
@@ -1129,8 +1129,8 @@ int32_t TPCCTables::orderSize(int64_t num_warehouses, int64_t num_transactions) 
     return ret;
 }
 
-int32_t TPCCTables::orderlineSize(int64_t num_warehouses, int64_t num_transactions) {
-    int32_t ret = 0;
+int64_t TPCCTables::orderlineSize(int64_t num_warehouses, int64_t num_transactions) {
+    int64_t ret = 0;
     int32_t krange = Order::INITIAL_ORDERS_PER_DISTRICT + num_transactions;
     for (int32_t i = 1; i <= num_warehouses; ++i) {
         for (int32_t j = 1; j <= District::NUM_PER_WAREHOUSE; ++j) {
@@ -1146,14 +1146,14 @@ int32_t TPCCTables::orderlineSize(int64_t num_warehouses, int64_t num_transactio
     return ret;
 }
 
-int32_t TPCCTables::newOrderSize() {
-    int32_t ret = 0;
+int64_t TPCCTables::newOrderSize() {
+    int64_t ret = 0;
     for (auto &no: neworders_) ret += no.second->size();
     return ret;
 }
 
-int32_t TPCCTables::historySize() {
-    int32_t ret = 0;
+int64_t TPCCTables::historySize() {
+    int64_t ret = 0;
     for (auto &h: history_) ret += h->size();
     return ret;
 }
